Refactor with subclasses
